Lunr and other changes

+++ Program.cs
@@ -1,6 +1,9 @@
-﻿using Markdig;
+﻿using System.Diagnostics;
+using AngryMonkey.Objects;
+using Markdig;
 using Markdig.Extensions.AutoIdentifiers;
 using Markdig.Syntax;
+using Newtonsoft.Json;
 using Spectre.Console;
 
 namespace AngryMonkey;
@@ -10,7 +13,8 @@ public static class Program
     //public static string source = "X:\\Gaea2\\Docs\\Source\\Nodes";
     //public static string destination = "X:\\Gaea2\\Docs\\staging";
 
-    public static string Root = @"X:\Docs\Gaea2-Docs\staging";
+    public static string RootFolder = @"X:\Docs\Gaea2-Docs";
+    public static string StagingFolder = $@"{RootFolder}\staging";
 
     public static string Templates = @"X:\Docs\Gaea2-Docs\template\";
     public static string HtmlTemplate = @"X:\Docs\Gaea2-Docs\template\template.html";
@@ -21,15 +25,27 @@ public static class Program
     private static Dictionary<string, string> Titles = [];
     private static Dictionary<string, string> mdTitles = [];
 
-    internal static Hive CurrentHive { get; set; }
+    internal static List<SearchObject> SearchObjects = [];
 
-    private const string SiteTitle = "Gaea Documentation";
+    internal static Hive CurrentHive { get; set; }
 
     public static void Main(string[] arguments)
     {// Synchronous
         AnsiConsole.Status()
             .Start("Monkey is getting angry...", ctx =>
             {
+                FileService.CopyDirectory($"{Templates}\\Assets", $"{StagingFolder}\\Assets");
+
+                if (Environment.CommandLine.Contains("--assets"))
+                {
+
+                    AnsiConsole.WriteLine();
+
+                    AnsiConsole.MarkupLine("[white]The Monkey copied assets only[/] [green][[Success]][/]");
+                    Environment.Exit(0);
+                    return;
+                }
+
                 pipeline = new MarkdownPipelineBuilder()
                     .UseAdvancedExtensions()
                     .UseAlertBlocks()
@@ -45,38 +61,91 @@ public static class Program
                     .UseCustomContainers()
                     .Build();
 
+                SearchObjects = [];
+
                 Hive[] hives =
                 [
                     new()
                     {
                         Name = "User Guide",
                         Source = @"X:\Docs\Gaea2-Docs\Source\UserGuide",
-                        Destination = @"X:\Docs\Gaea2-Docs\staging\Manual",
+                        Destination = @"X:\Docs\Gaea2-Docs\staging\",
+                        ShortName = "Manual"
                     },
                     new()
                     {
                         Name = "Node Reference",
                         Source = @"X:\Docs\Gaea2-Docs\Source\Nodes",
-                        Destination = @"X:\Docs\Gaea2-Docs\staging\Nodes",
+                        Destination = @"X:\Docs\Gaea2-Docs\staging\Reference",
+                        ShortName = "Reference",
+                        URL = "/Reference"
                     }
                 ];
 
                 Html = File.ReadAllText(HtmlTemplate);
 
-                FileService.CopyDirectory($"{Templates}\\Assets", $"{Root}\\Assets");
+                //Directory.Delete(Root, true);
+                //Directory.CreateDirectory(Root);
 
                 foreach (Hive hive in hives)
                 {
                     ctx.Status($"Processing {hive.Name}...");
                     CurrentHive = hive;
+                    //RenameFiles(hive);
+                    //PreprocessMarkdown(hive);
                     ParseTOC(hive);
-                    // PreprocessMarkdown();
                     RecreateDestination(hive);
                     ProcessMarkdown(hive);
                 }
+
+                File.WriteAllText($"{StagingFolder}\\search.json", JsonConvert.SerializeObject(SearchObjects, new JsonSerializerSettings { Formatting = Formatting.None }));
+
+                //ctx.Status("Running PageFind...");
+
+                //Process.Start(new ProcessStartInfo("PageFind", @"--site .\staging\ --quiet")
+                //{
+                //    WorkingDirectory = RootFolder
+                //}).WaitForExit();
             });
 
-        Console.WriteLine("The Monkey has gone home [Success]");
+        AnsiConsole.WriteLine();
+
+        AnsiConsole.MarkupLine("[white]The Monkey has gone home[/] [green][[Success]][/]");
+
+        //File.WriteAllLines($"{Root}\\images.txt", images);
+    }
+
+    private static void RenameFiles(Hive hive)
+    {
+        string tocFile = Path.Combine(hive.Source, "SUMMARY.md");
+
+        string[] lines = File.ReadAllLines(tocFile);
+
+        int counter = 1;
+
+        string lastPath = "";
+
+        foreach (string line in lines)
+        {
+            if (!line.Contains(".md"))
+            {
+                counter = 1;
+                continue;
+            }
+
+            string file = line.Split('(')[1].TrimEnd(')');
+            string fullPath = Path.Combine(hive.Source, file.Replace('/', '\\'));
+            string filePath = Path.GetDirectoryName(fullPath);
+
+            if (filePath != lastPath)
+            {
+                lastPath = filePath;
+                counter = 1;
+            }
+
+            File.Move(fullPath, $"{filePath}\\{counter:00}-{Path.GetFileName(fullPath)}");
+            counter++;
+        }
     }
 
     private static void PreprocessMarkdown(Hive hive)
@@ -84,31 +153,17 @@ public static class Program
         string[] md = Directory.GetFiles(hive.Source, "*.md", SearchOption.AllDirectories);
         foreach (var file in md)
         {
-            Console.WriteLine(file);
-            string content = File.ReadAllText(file);
-            content = content
-                .Replace("{% hint style=\"warning\" %}", ":::warning")
-                .Replace("{% hint style=\"danger\" %}", ":::danger")
-                .Replace("{% hint style=\"note\" %}", ":::note")
-                .Replace("{% hint style=\"tip\" %}", ":::tip")
-                .Replace("{% hint style=\"success\" %}", ":::success")
-                .Replace("{% hint style=\"info\" %}", ":::info")
-                .Replace("{% endhint %}", ":::")
-                .Replace(".md \"mention\"", ".html");
-
-            string title = "untitled";
-            try
+            // Console.WriteLine(file);
+            string[] content = File.ReadAllLines(file);
+            var dic = FrontMatter.GetFrontMatter(content);
+            if (!dic.ContainsKey("uid"))
             {
-                title = TOC[file.Replace(hive.Source, "").Replace("\\", "/").TrimStart('/')];
+                if (dic.ContainsKey("title"))
+                { dic.Add("uid", FrontMatter.Slugify(dic["title"])); }
+                else
+                { dic.Add("uid", Path.GetFileNameWithoutExtension(file)); }
             }
-            catch (Exception)
-            {
-                title = "untitled";
-            }
-
-            content = FrontMatter.EnsureYamlTitle(content, title);
-
-            File.WriteAllText(file, content);
+            File.WriteAllText(file, FrontMatter.ReplaceFrontMatter(content, dic));
         }
     }
 
@@ -119,6 +174,7 @@ public static class Program
         mdTitles.Clear();
 
         string tocFile = Path.Combine(hive.Source, "SUMMARY.md");
+        string hivePrefix = hive.Destination.Split('\\')[^1];
 
         string[] lines = File.ReadAllLines(tocFile);
         foreach (var line in lines)
@@ -132,7 +188,7 @@ public static class Program
                 if (startTitle >= 0 && endTitle > startTitle && startLink >= 0 && endLink > startLink)
                 {
                     string title = line.Substring(startTitle, endTitle - startTitle);
-                    string link = line.Substring(startLink, endLink - startLink);
+                    string link = $"{hivePrefix}/{line.Substring(startLink, endLink - startLink)}";
                     TOC[link] = title;
                     Titles[title] = link;
                 }
@@ -140,12 +196,15 @@ public static class Program
         }
     }
 
-    public static void RecreateDestination(Hive hive)
+    public static void RecreateDestination(Hive hive, bool delete = false)
     {
-        // delete all contents in the destination directory
-        if (Directory.Exists(hive.Destination))
+        if (delete)
         {
-            Directory.Delete(hive.Destination, true);
+            // delete all contents in the destination directory
+            if (Directory.Exists(hive.Destination))
+            {
+                Directory.Delete(hive.Destination, true);
+            }
         }
 
         FileService.CopyDirectory(hive.Source, hive.Destination);
@@ -162,26 +221,28 @@ public static class Program
         //}
     }
 
+    private static List<string> images = new List<string>(5000);
+
     public static void ProcessMarkdown(Hive hive)
     {
         string[] md = Directory.GetFiles(hive.Destination, "*.md", SearchOption.AllDirectories);
 
         var resolver = new MarkdownTitleResolver(hive.Destination);
-        var toc = SummaryToc.BuildFromFile(summaryMdPath: hive.Destination + "\\summary.md",
+        List<TocNode> toc = SummaryToc.BuildFromFile(summaryMdPath: hive.Destination + "\\summary.md",
             options: new SummaryTocOptions
             {
-                BaseUrl = "/",
+                BaseUrl = hive.URL ?? "/",
                 PromoteParentLinksToGroups = true,
                 OverviewTitle = "Overview"
             });
 
-        var js = SummaryToc.ToJavascript(toc);
+        string js = SummaryToc.ToJavascript(toc);
+        Directory.CreateDirectory($"{StagingFolder}\\assets");
+        File.WriteAllText($@"{StagingFolder}\assets\TOC_{hive.ShortName}.js", js);
 
         foreach (var file in md)
         {
             string html = Html;
-
-            string hiveName = hive.Name ?? "NO_HIVE";
 
             string content = File.ReadAllText(file).Replace(".md", ".html");
 
@@ -189,19 +250,26 @@ public static class Program
 
             MarkdownLinkTextToTargetTitle.Rewrite(doc, file, resolver);
 
-            html = html.Replace("%%CONTENT%%", doc.ToHtml(pipeline));
+            string contentHTML = doc.ToHtml(pipeline);
+            html = html.Replace("%%CONTENT%%", contentHTML);
+
+            //images.AddRange(IMG.ExtractImageUrlsFromHtmlAsync(contentHTML).Result);
 
             try
             {
-                string title = resolver.TryGetTitle(file) ?? "Untitled";
+                var dic = FrontMatter.GetFrontMatter(content.Split(Environment.NewLine));
+                string title = dic["title"];
 
-                var nav = DocNav.Build(file.Replace(hive.Destination, ""), hiveName, title);
+                var nav = DocNav.Build(file.Replace(hive.Destination, ""), hive.Name, title);
 
-                html = html.Replace("%%TITLE%%", $"{title} - {SiteTitle}")
-                    .Replace("%%HIVE%%", hiveName)
-                    .Replace("//%%TOC%%", js)
+                html = html.Replace("%%TITLE%%", title)
+                    .Replace("%%HIVE%%", hive.Name)
                     .Replace("%%CRUMBS%%", $"{nav.BreadcrumbHtml}")
-                    .Replace("%%PAGETITLE%%", $"\n{nav.PretitleHtml}\n{nav.PageTitleHtml}\n<hr>");
+                    .Replace("%%SHORTNAME%%", hive.ShortName)
+                    .Replace("%%SLUG%%", dic["uid"])
+                    .Replace("%%PAGETITLE%%", $"\n{nav.PretitleHtml}\n{nav.PageTitleHtml}");
+
+                SearchObjects.Add(SearchObject.ToSearchObject(content, title, hive, dic["uid"]));
             }
             catch (Exception ex)
             {
